requirements.txt
asgiref==3.9.1
black==25.1.0
click==8.2.1
Django==5.2.6
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
psycopg2==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3



